+++ iGCS/UserInterface/ViewControllers/WaypointMapBaseController.h
@@ -22,7 +22,7 @@
 @interface WaypointMapBaseController : UIViewController <MKMapViewDelegate, CLLocationManagerDelegate> {
 @private
     MKPolyline *_waypointRoutePolyline;
-    int _currentWaypointNum;
+    NSInteger _currentWaypointNum;
 
     MKPolyline *_trackPolyline;
     MKMapPoint *_trackMKMapPoints;
@@ -39,9 +39,9 @@
 @property (nonatomic, retain) IBOutlet MKMapView *mapView;
 
 - (void) removeExistingWaypointAnnotations;
-- (WaypointAnnotation *) getWaypointAnnotation:(int)waypointSeq;
+- (WaypointAnnotation *) getWaypointAnnotation:(NSInteger)waypointSeq;
 - (void) resetWaypoints:(WaypointsHolder *)waypoints;
-- (void) maybeUpdateCurrentWaypoint:(int)newCurrentWaypointSeq;
+- (void) maybeUpdateCurrentWaypoint:(NSInteger)newCurrentWaypointSeq;
 
 - (void) makeWaypointsDraggable:(BOOL)_draggableWaypointsP;
 - (NSString*) waypointNumberForAnnotationView:(mavlink_mission_item_t)item;
@@ -49,7 +49,7 @@
 - (void) addToTrack:(CLLocationCoordinate2D)pos;
 
 // Following methods are intended to be overridden by subclasses
-- (void) waypointWithSeq:(int)waypointSeq wasMovedToLat:(double)latitude andLong:(double)longitude;
+- (void) waypointWithSeq:(NSInteger)waypointSeq wasMovedToLat:(double)latitude andLong:(double)longitude;
 - (void) customizeWaypointAnnotationView:(MKAnnotationView*)view;
 - (void) handleLongPressGesture:(UIGestureRecognizer*)sender;
